refactored staff module 2

+++ src/staff/staff.service.ts
@@ -2,30 +2,24 @@ import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateStaffDto } from './dto/create-staff.dto';
 import { UpdateStaffDto } from './dto/update-staff.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { Role } from '@prisma/client';
 import { hash } from 'argon2';
-
-type FindAllWhere = {
-  role?: {
-    not: Role;
-  };
-  name?: {
-    startsWith: string;
-    mode: 'insensitive';
-  };
-  surname?: {
-    startsWith: string;
-    mode: 'insensitive';
-  };
-  specializationId?: string;
-  specialization?: {
-    isNot: null;
-  };
-};
+import { FindAllStaffDto } from './dto/find-all-staff.dto';
+import { AvatarsService } from 'src/avatars/avatar.service';
 
 @Injectable()
 export class StaffService {
-  constructor(private prisma: PrismaService) {}
+  constructor(
+    private prisma: PrismaService,
+    private readonly avatarsService: AvatarsService,
+  ) {}
+
+  private parseBirthday(birthday?: string) {
+    return birthday && new Date(birthday);
+  }
+
+  private async hashPassword(password?: string) {
+    return password && (await hash(password));
+  }
 
   async create(createStaffDto: CreateStaffDto) {
     const user = await this.findOneByEmail(createStaffDto.email);
@@ -34,14 +28,17 @@ export class StaffService {
       throw new BadRequestException('Staff member already exists');
     }
 
+    const hashedPassword = await this.hashPassword(createStaffDto.password);
+    const birthday = this.parseBirthday(createStaffDto.birthday);
+
     const { id } = await this.prisma.staff.create({
       data: {
         user: {
           create: {
             ...createStaffDto,
             isEmailVerified: true,
-            password: await hash(createStaffDto.password),
-            birthday: new Date(createStaffDto.birthday),
+            password: hashedPassword,
+            birthday: birthday,
           },
         },
       },
@@ -49,69 +46,61 @@ export class StaffService {
     return id;
   }
 
-  findAll(
-    role: Role,
-    specializationId?: string,
-    surname?: string,
-    name?: string,
-  ) {
-    const where: FindAllWhere = {};
-    if (role !== Role.Admin) {
-      where.role = {
-        not: Role.Admin,
-      };
-      where.specialization = {
-        isNot: null,
-      };
-    }
-    if (specializationId) {
-      where.specializationId = specializationId;
-    }
-    if (surname) {
-      where.surname = {
-        startsWith: surname,
-        mode: 'insensitive',
-      };
-    }
-    if (name) {
-      where.name = {
-        startsWith: name,
-        mode: 'insensitive',
-      };
-    }
-
-    const select = {
-      id: true,
-      name: true,
-      surname: true,
-      experience: true,
-      gender: true,
-      specialization: {
-        select: {
-          title: true,
+  async findAll({
+    name,
+    surname,
+    specializationId,
+    page,
+    take,
+  }: FindAllStaffDto) {
+    const conditions = {
+      user: {
+        name: {
+          startsWith: name || '',
         },
+        surname: {
+          startsWith: surname || '',
+        },
+        specializationId,
       },
     };
 
-    return this.prisma.staff.findMany({
-      where: where,
-      select: select,
-    });
-  }
-
-  findAllAdmins() {
-    return this.prisma.staff.findMany({
-      where: {
-        role: 'Admin',
-      },
-      include: {
-        user: {
-          select: {
-            id: true,
+    const [staff, totalCount] = await this.prisma.$transaction([
+      this.prisma.staff.findMany({
+        where: conditions,
+        select: {
+          id: true,
+          user: {
+            select: {
+              name: true,
+              surname: true,
+              gender: true,
+            },
+          },
+          experience: true,
+          specialization: {
+            select: {
+              title: true,
+            },
           },
         },
+        skip: (page - 1) * take,
+        take: take,
+      }),
+      this.prisma.staff.count({ where: conditions }),
+    ]);
+
+    return {
+      data: staff.map((staffMember) => ({
+        ...staffMember.user,
+        id: staffMember.id,
+      })),
+      pagination: {
+        page,
+        take,
+        total: totalCount,
       },
-    });
+    };
   }
 
   async findOneById(id: string) {
@@ -138,85 +127,6 @@ export class StaffService {
     return result;
   }
 
-  async getAvailableTime(staffId: string, startDate: string, endDate: string) {
-    const startDateParsed = new Date(startDate);
-    const endDateParsed = new Date(endDate);
-
-    const { schedule, scheduleExceptions, appointments } =
-      await this.prisma.staff.findUnique({
-        where: {
-          id: staffId,
-        },
-        select: {
-          schedule: {
-            where: {
-              dayOfWeek: {
-                gte: startDateParsed.getDay(),
-                lte: endDateParsed.getDay(),
-              },
-            },
-          },
-          scheduleExceptions: true,
-          appointments: {
-            where: {
-              dateTime: {
-                gte: startDateParsed,
-                lte: endDateParsed,
-              },
-            },
-          },
-        },
-      });
-
-    const availableTime: { date: string; availableSlots: number[] }[] = [];
-    const bookedHoursMap = new Map<string, Set<number>>();
-
-    appointments.forEach((appointment) => {
-      const dateKey = appointment.dateTime.toISOString().split('T')[0];
-      const hour = appointment.dateTime.getHours();
-
-      if (!bookedHoursMap.has(dateKey)) {
-        bookedHoursMap.set(dateKey, new Set());
-      }
-      bookedHoursMap.get(dateKey)?.add(hour);
-    });
-
-    for (
-      let d = new Date(startDateParsed);
-      d <= endDateParsed;
-      d.setDate(d.getDate() + 1)
-    ) {
-      const dayOfWeek = d.getDay();
-      const dateStr = d.toISOString().split('T')[0];
-
-      const exception = scheduleExceptions.find(
-        (ex) => ex.date.toISOString().split('T')[0] === dateStr,
-      );
-      if (exception) {
-        continue;
-      }
-
-      const daySchedule = schedule.find((sch) => sch.dayOfWeek === dayOfWeek);
-
-      if (daySchedule && daySchedule.from !== null && daySchedule.to !== null) {
-        const availableSlots: number[] = [];
-        const bookedHours = bookedHoursMap.get(dateStr) || new Set();
-
-        for (let hour = daySchedule.from; hour < daySchedule.to; hour++) {
-          if (!bookedHours.has(hour)) {
-            availableSlots.push(hour);
-          }
-        }
-
-        if (availableSlots.length > 0) {
-          availableTime.push({ date: dateStr, availableSlots });
-        }
-      }
-    }
-
-    return availableTime;
-  }
-
   async findOneByEmail(email: string) {
     return await this.prisma.user.findUnique({
       where: {
@@ -232,10 +142,11 @@ export class StaffService {
   }
 
   async update(id: string, updateStaffDto: UpdateStaffDto) {
-    const data = { ...updateStaffDto };
-    if (updateStaffDto.birthday) {
-      data.birthday = new Date(updateStaffDto.birthday);
-    }
+    const data = {
+      ...updateStaffDto,
+      birthday: this.parseBirthday(updateStaffDto.birthday),
+      password: await this.hashPassword(updateStaffDto.password),
+    };
 
     const updateUser = await this.prisma.staff.update({
       where: {
@@ -262,14 +173,19 @@ export class StaffService {
   }
 
   async remove(id: string) {
+    const staffMember = await this.prisma.staff.findUnique({
+      where: { id },
+      select: { userId: true },
+    });
+
+    if (!staffMember) {
+      throw new BadRequestException('Staff member not found');
+    }
+
+    await this.avatarsService.deleteByUserId(id);
     await this.prisma.$transaction(async (prisma) => {
-      const deletedEmployee = await prisma.staff.delete({
-        where: { id },
-        select: { userId: true },
-      });
-      await prisma.user.delete({
-        where: { id: deletedEmployee.userId },
-      });
+      await prisma.staff.delete({ where: { id } });
+      await prisma.user.delete({ where: { id: staffMember.userId } });
     });
   }
 }
